PlanB
This is really a bunch of stuff that will most likely get thrown out. It's the android build folder for PlanB (localized storage). Endgame looks like this: we pull the java files for local storage and throw it into Dannys cordova project so that the APK he builds has the ability to do localized storage. The reason I'm doing the build folder is because it makes everything easier and I dont want to risk not storing a file we would later need.